pyproject.toml
[tool.poetry]
name = "aw-watcher-quattrod"
version = "0.2.0"
description = "Window watcher for ActivityWatch"
authors = ["Erik Bjäreholt <erik@bjareho.lt>"]
license = "MPL-2.0"

[tool.poetry.scripts]
aw-watcher-quattrod = "aw_watcher_quattrod:main"

[tool.poetry.dependencies]
python = "^3.8,<3.13"
aw-client = "^0.5"
pywin32 = {version = "306", platform = "win32"}
wmi = {version = "*", platform = "win32"}
pyobjc-framework-ApplicationServices = { version = "*", platform="darwin"}
pyobjc-framework-CoreText = {version = "*", platform="darwin"}
pyobjc-framework-OSAKit = {version = "*", platform="darwin"}

# locked due to https://github.com/python-xlib/python-xlib/pull/242 leading to 100% CPU stalls
# see: https://github.com/ActivityWatch/aw-watcher-window/issues/89
# also locked in aw-watcher-afk
python-xlib = {version = "0.31", platform = "linux"}
psutil = "^5.9.7"
pywinauto = "^0.6.8"
pyqt6 = "^6.6.1"
flask = "^3.0.1"
toml = "^0.10.2"
python-dateutil = "^2.9.0.post0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
mypy = "*"
macholib = {version = "*", platform = "darwin"}  # Needed for pyinstaller
pyinstaller = {version = "*"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
